+++ apps/web/src/core/hooks/index.ts
@@ -1,0 +1,4 @@
+// Core app hooks exports
+export * from './useStoreSelectors';
+export * from './useHistoryActions';
+export * from './useExecutionMonitor';
