+++ apps/web/src/core/hooks/useHistoryActions.ts
@@ -1,6 +1,6 @@
 // Hook for history (undo/redo) actions
 import { useCallback } from 'react';
-import { useHistoryStore, useConsolidatedDiagramStore } from '@/shared/stores';
+import { useHistoryStore, useConsolidatedDiagramStore } from '@/core/stores';
 
 export const useHistoryActions = () => {
   const { undo, redo, canUndo, canRedo, saveToHistory, clearHistory, initializeHistory } = useHistoryStore();
