+++ apps/web/src/core/index.ts
@@ -1,0 +1,10 @@
+// Core app infrastructure exports
+
+// Stores
+export * from './stores';
+
+// Contexts  
+export * from './contexts/DiagramContext';
+
+// Core hooks
+export * from './hooks';
