+++ apps/web/src/core/stores/consolidatedDiagramStore.ts
@@ -10,7 +10,7 @@ import {
   getReactFlowType, OnArrowsChange, Arrow, applyArrowChanges, addArrow,
   StartBlockData, PersonJobBlockData, DBBlockData, JobBlockData,
   ConditionBlockData, EndpointBlockData, createErrorHandlerFactory
-} from '../types';
+} from '@/shared/types';
 import { sanitizeDiagram } from "@/features/serialization/utils/diagramSanitizer";
 import { createPersonCrudActions, createApiKeyCrudActions } from "@/shared/utils/storeCrudUtils";
 import { API_ENDPOINTS, getApiUrl } from '@/shared/utils/apiConfig';
