+++ apps/web/src/core/stores/consolidatedUIStore.ts
