+++ apps/web/src/features/canvas/components/MemoryLayer.tsx
@@ -1,6 +1,6 @@
 // apps/web/src/components/diagram/MemoryLayer.tsx
 import React from 'react';
-import { useConsolidatedDiagramStore, useExecutionStore } from '@/shared/stores';
+import { useConsolidatedDiagramStore, useExecutionStore } from '@/core/stores';
 import { Users, MessageSquare, Database } from 'lucide-react';
 
 const MemoryLayer: React.FC = () => {
