+++ apps/web/src/features/conversation/components/ConversationDashboard.tsx
@@ -6,7 +6,7 @@ import {
 import { Button, Input, Select, SelectItem } from '@/shared/components';
 import { downloadJson } from '@/shared/utils/downloadUtils';
 import { toast } from 'sonner';
-import { usePersons, useSelectedElement, useExecutionStatus } from '@/shared/hooks/useStoreSelectors';
+import { usePersons, useSelectedElement, useExecutionStatus } from '@/core/hooks/useStoreSelectors';
 import { useConversationData, useMessagePolling } from '../hooks';
 import { MessageList } from './MessageList';
 import { ConversationMessage, ConversationFilters } from '../types';
