--- apps/web/src/features/diagram/hooks/index.ts
@@ -1,5 +1,0 @@
-export { useDiagramActions } from './useDiagramActions';
-// The following hooks have been moved:
-// useDiagramRunner -> @/features/execution/hooks/useDiagramRunner
-// useFileImport -> @/features/serialization/hooks/useFileImport
-// useHybridExecution -> @/features/execution/hooks/useHybridExecution
