--- apps/web/src/features/diagram/index.ts
@@ -1,19 +1,0 @@
-// Components - moved to features/canvas
-// export { default as Canvas } from './components/Canvas';
-// export { default as MemoryLayer } from './components/MemoryLayer';
-
-// Diagram UI wrappers - moved to features/canvas/wrappers
-// export { CustomArrow, ContextMenu } from './wrappers';
-
-// Hooks
-export * from './hooks/useDiagramActions';
-// The following hooks have been moved:
-// useDiagramRunner -> features/execution/hooks
-// useFileImport -> features/serialization/hooks
-
-// Utils
-export * from './utils/diagramHelpers';
-// The following utils have been moved:
-// diagramSanitizer -> features/serialization/utils
-// yamlExporter -> features/serialization/converters
-// canvasUtils -> features/canvas/utils
