+++ apps/web/src/features/execution/hooks/useDiagramRunner.ts
@@ -1,5 +1,5 @@
 import { useState, useCallback, useRef, useEffect } from 'react';
-import { useConsolidatedDiagramStore, useExecutionStore } from '@/shared/stores';
+import { useConsolidatedDiagramStore, useExecutionStore } from '@/core/stores';
 import { toast } from 'sonner';
 import { createErrorHandlerFactory, PersonDefinition } from '@/shared/types';
 import { API_ENDPOINTS, getApiUrl, getStreamingUrl } from '@/shared/utils/apiConfig';
