+++ apps/web/src/shared/index.ts
@@ -1,14 +1,12 @@
-export * from './hooks/useStoreSelectors';
-export * from './hooks/useHistoryActions';
-export * from './hooks/useExecutionMonitor';
-
+// Only export truly shared utilities (components, utils, types)
 export * from './components/common/FileUploadButton';
 export * from './components/skeletons/SkeletonComponents';
 
-export * from './contexts/DiagramContext';
+// export * from './hooks/useDownload'; // Moved to features/serialization/hooks/
 
 export * from './utils/apiConfig';
 export * from './utils/constants';
+export * from './utils/diagramUtils';
 export * from './utils/downloadUtils';
 export * from './utils/promptUtils';
 export * from './utils/storeCrudUtils';
