+++ apps/web/src/shared/utils/diagramUtils.ts
